requirements.txt
rich
requests
psutil
python-whois
